Consolidated configs, got rid of model class
