+++ src/models/layers.py
@@ -32,19 +32,3 @@ def resnet(input, params):
         for _ in range(params['num_rb']):
             x = resnet_block(x, params)
     return x
-
-
-def optimize(loss, params):
-    if params['optimizer_type'] == 'rmsprop':
-        optimizer = tf.train.RMSPropOptimizer(params["learning_rate"])
-    elif params['optimizer_type'] == 'sgd':
-        optimizer = tf.train.GradientDescentOptimizer(params["learning_rate"])
-    elif params['optimizer_type'] == 'adam':
-        optimizer = tf.train.AdamOptimizer(params["learning_rate"])
-    else:
-        raise Exception('Unkown optimizer type: %s' % params['optimizer_type'])
-    # Add mean and variance ops to dependencies so batch norm works during training
-    update_ops = tf.get_collection(tf.GraphKeys.UPDATE_OPS, scope=params["model_name"])
-    with tf.control_dependencies(update_ops):
-        train_op = optimizer.minimize(loss, global_step=tf.train.get_global_step())
-    return train_op
